feat: implement authentication and authorization in each method

+++ src/main/java/org/example/microTech/controllers/ClientController.java
@@ -1,10 +1,15 @@
 package org.example.microTech.controllers;
 
 
+import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import org.example.microTech.dto.*;
 
+import org.example.microTech.entities.User;
+import org.example.microTech.enums.UserRole;
+import org.example.microTech.exceptions.ForbiddenException;
+import org.example.microTech.exceptions.UnauthorizedException;
 import org.example.microTech.services.ClientService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -21,7 +26,10 @@ public class ClientController {
 
 
     @PostMapping
-    public ResponseEntity<ApiResponse> createClient(@Valid @RequestBody ClientCreateDTO dto){
+    public ResponseEntity<ApiResponse> createClient(@Valid @RequestBody ClientCreateDTO dto, HttpSession session) {
+        User user = (User) session.getAttribute("user");
+        if (user == null) throw new UnauthorizedException("You must login");
+        if (!user.getRole().equals(UserRole.ADMIN)) throw new ForbiddenException("Access denied");
         ClientResponseDTO clientResponseDTO = clientService.createClient(dto);
         ApiResponse response =  ApiResponse.builder().message("Client created successfully!")
                 .data(clientResponseDTO)
@@ -31,7 +39,10 @@ public class ClientController {
     }
 
     @GetMapping
-    public ResponseEntity<ApiResponse> getAllClients(){
+    public ResponseEntity<ApiResponse> getAllClients(HttpSession session) {
+        User user = (User) session.getAttribute("user");
+        if (user == null) throw new UnauthorizedException("You must login");
+        if (!user.getRole().equals(UserRole.ADMIN)) throw new ForbiddenException("Access denied");
         List<ClientResponseDTO> clients = clientService.getAllClients();
         ApiResponse response = ApiResponse.builder()
                 .status(HttpStatus.OK.value())
@@ -41,7 +52,9 @@ public class ClientController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<ApiResponse> getById(@PathVariable Long id) {
+    public ResponseEntity<ApiResponse> getById(@PathVariable Long id, HttpSession session) {
+        User user = (User) session.getAttribute("user");
+        if (user == null) throw new UnauthorizedException("You must login");
         ClientResponseDTO client = clientService.getClientById(id);
         ApiResponse response = ApiResponse.builder()
                 .message("Successfully retrieved client")
@@ -52,8 +65,10 @@ public class ClientController {
     }
 
     @PutMapping("/{id}")
-    public  ResponseEntity<ApiResponse>update(@PathVariable Long id, @RequestBody @Valid ClientUpdateDTO dto) {
-
+    public  ResponseEntity<ApiResponse>update(@PathVariable Long id, @RequestBody @Valid ClientUpdateDTO dto, HttpSession session) {
+        User user = (User) session.getAttribute("user");
+        if (user == null) throw new UnauthorizedException("You must login");
+        if (!user.getRole().equals(UserRole.ADMIN)) throw new ForbiddenException("Access denied");
         ClientResponseDTO client = clientService.updateClient(id, dto);
         ApiResponse response = ApiResponse.builder()
                 .message("Successfully updated clinet")
@@ -64,7 +79,10 @@ public class ClientController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<ApiResponse> deleteClient(@PathVariable Long id) {
+    public ResponseEntity<ApiResponse> deleteClient(@PathVariable Long id,HttpSession session) {
+        User user = (User) session.getAttribute("user");
+        if (user == null) throw new UnauthorizedException("You must login");
+        if (!user.getRole().equals(UserRole.ADMIN)) throw new ForbiddenException("Access denied");
         ClientResponseDTO client = clientService.deleteClient(id);
         ApiResponse response = ApiResponse.builder()
                 .message("Supplier deleted successfully!")
